fix canonical form for seqinr url
This was the only instance in which the PDF manual for seqinr was linked, but not the first time it was referenced. While the link does properly re-direct, I wouldn't be surprised if the CRAN team complains. I removed the seqinr.pdf part of the URL to ensure smooth sailing.